feat(post): add new blog post about becoming an Estonia e-resident
This commit introduces a new post in the Life category regarding the process of becoming an e-resident of Estonia. The post describes the writer's experience of applying for and becoming a digital citizen, and discusses the potential benefits and opportunities that come with e-residency. It also provides insights for those considering to follow a similar path.