feat(post): add Medium-style image zoom feature
A post was added detailing the implementation of a Medium-style image zoom feature in v1.9.0 release. The feature allows users to focus on an image by clicking or tapping on it, causing it to expand to the full screen. This feature was initially implemented using Framer Motion in v1.8.0 release, but due to slight flickering issues in non-Chrome browsers, it was temporarily disabled. The version 1.9.0 implements this feature again with the usage of the react-medium-image-zoom library, providing a smoother and higher-quality user experience.